Code refactor
This commit contains changes to the location of
several function definitions and declarations in
an attempt to create an interface to be used by
other database systems:

- Functions: read_buffer_pipe(), execve_binary() definitions
  were moved to util.h and declarations to util.c

- Macro: EMAIL_BODY_LENGTH moved to util.h

- Cleanup of `include` libraries in postgres.c
